requirements.txt
torch>=2.0.0
# torchvision: required for satellite V_Dataset / build_loaders; CIFAR ablations use utils/cifar10_standalone (no torchvision)
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tqdm>=4.65.0
Pillow>=9.4.0
rasterio>=1.3.0
tifffile>=2023.1.23
opencv-python>=4.7.0
einops>=0.6.0
albumentations>=1.3.0
pyyaml>=6.0
tensorboard>=2.14.0